Adding comments and fixing sinc_resample test
