+++ include/phaseshapers.h
@@ -17,9 +17,11 @@
 namespace dsp
 {
 
+/// @brief Phaseshaper oscillator that can morph between different waveforms
 class Phaseshaper
 {
   public:
+    /// @brief Waveform types
     enum class Waveform : uint8_t
     {
         VARIABLE_SLOPE = 0,
@@ -34,15 +36,14 @@ class Phaseshaper
     Phaseshaper() = default;
     ~Phaseshaper() = default;
 
-    void Init(float sampleRate)
-    {
-        m_sampleRate = sampleRate;
-        m_phase = 0.f;
-        m_freq = 220.f;
-        m_phaseIncrement = m_freq / m_sampleRate;
-        m_period = m_sampleRate / m_freq;
-    }
+    /// @brief Initializes the oscillator
+    /// @param sampleRate
+    void Init(float sampleRate);
 
+    /// @brief Sets the waveform
+    /// @param wave floating point value between Waveform::VARIABLE_SLOPE and Waveform::NUM_WAVES
+    /// @note The fractional part of the value is used to interpolate between waveforms.
+    /// For example, a value of 1.5 will interpolate between Waveform::SOFTSYNC and Waveform::WAVESLICE equally.
     void SetWaveform(float wave)
     {
         m_waveform = wave;
@@ -55,50 +56,24 @@ class Phaseshaper
         m_period = m_sampleRate / m_freq;
     }
 
-    void SetMod(float mod)
-    {
-        m_mod = mod;
-    }
+    /// @brief Sets the modulation amount
+    /// @param mod Modulation amount, clamped between 0 and 1
+    /// The modulation parameter is used differently depending on the waveform.
+    /// For example, for Waveform::VARIABLE_SLOPE, the modulation parameter controls the width of the waveform.
+    void SetMod(float mod);
 
     float Process();
 
   private:
-    float ProcessWaveSlice();
-    float ProcessHardSync();
-    float ProcessSoftSync();
-    float ProcessTriMod();
-    float ProcessSupersaw();
-    float ProcessVarSlope();
-    float ProcessVarTri();
+    float ProcessWaveSlice() const;
+    float ProcessHardSync() const;
+    float ProcessSoftSync() const;
+    float ProcessTriMod() const;
+    float ProcessSupersaw() const;
+    float ProcessVarSlope() const;
+    float ProcessVarTri() const;
 
-    inline float ProcessWave(Waveform wave)
-    {
-        float out = 0.f;
-        switch (wave)
-        {
-        case Waveform::VARIABLE_SLOPE:
-            out = ProcessVarSlope();
-            break;
-        case Waveform::SOFTSYNC:
-            out = ProcessSoftSync();
-            break;
-        case Waveform::WAVESLICE:
-            out = ProcessWaveSlice();
-            break;
-        case Waveform::SUPERSAW:
-            out = ProcessSupersaw();
-            break;
-        case Waveform::HARDSYNC:
-            out = ProcessHardSync();
-            break;
-        case Waveform::TRIANGLE_MOD:
-            out = ProcessTriMod();
-            break;
-        default:
-            break;
-        }
-        return out;
-    }
+    float ProcessWave(Waveform wave) const;
 
   private:
     float m_sampleRate = 0.f;
